requirements.txt
python-telegram-bot[job-queue]>=20.0
pycups>=2.0.0
Pillow>=10.0
python-dotenv>=1.0
google-genai>=1.0
